Don't escape extension tags shadowing html5 tags if found in source
Follow up to 3626487

If a closing extension tag isn't found, Parsoid will continue with the
tag as if it's an html tag and rely on the sanitizer to escape tags
which aren't allowed literals.  See Parsoid commit adeb4e4 and the
equivalent core commit 674e838.

So far, <pre> is the only tag that's both an extension tag and an
allowed literal.  Support roundtripping tags found in source.

Bug: T300436
Change-Id: I2636ae48fe9754ccb9b538e9f884f62b2dee1a08