SiteConfig: Clear extension cache on extension registration
This ensures that subsequent fetches of extension configs picks
up the extension registration.

In production settings, this is unlikely to be of much concern
since extensions are expected to be processed as part of startup.

This is only useful in parser test runs to enable reuse of a shared
SiteConfig object (via a shared Parsoid object or otherwise). But,
it can be hacked around by calling unregisterExtensionModule on an
invalid ext id (which clears the extension cache).

Change-Id: Ia08e657b7812ff432aaa461840b8258f0e2dc150