Add forward compatibility to avoid serializing alt from caption
In I9d82c8003fd67cd984a8f4523e4993ed1f22b5d2, Parsoid will start
duplicating the caption to the alt attribute if media haven't specified
an explicit alt option and if the caption isn't visible.  This matches
the legacy parser's behaviour.

However, Parsoid needs to be able to roundtrip this and now the alt
attribute can either originate from the explicit option or from a
caption and it needs to distinguish the two to avoid dirty diffs.

The simplest thing to do is just drop the alt option if it's equivalent
to the caption.  That's backwards compatible since, at least for the
legacy parser, the alt would have been added from the caption if it
wasn't present and they would be the same content.

There are a few hitches though:

* Any slight deviation could result in duplicating the caption into the
  alt option when serializing.  For example, since ref processing
  happens after adding media info, the linkbacks aren't available in the
  textContent added to the alt.  However, when serializing, they are in
  the caption elements.  So, a special handler is added in this patch to
  drop the linkbacks for the purpose of comparison.

* VE will display the content in both the alt field and the caption, and
  they'd need to be edited in sync to avoid being duplicated and diverge.
  The editing experience when the caption isn't visible can stand to be
  improved as it is though.

Bug: T297443
Bug: T162360
Bug: T63566
Change-Id: I2bf43c7e83283f43e047229eb53c244918fcbb0c