Use caption as alt on imgs when not present and caption isn't visible
Similar to Icbc36b6e9aa1b9f4f27c23f4833c626a725cc154 but we need to be
able to roundtrip the alt attribute if the |alt= media option is set.
Serialization was added for forward compatibility in
I2bf43c7e83283f43e047229eb53c244918fcbb0c.

This matches the legacy parser's behaviour.  However, in Parsoid, a
caption not being visible doesn't always mean it isn't accessible to a
screen reader.  Some block media have a figcaption but set "display:
none" on it.  So, if this was only about accessibility, it would only
need to be applied to inline images.  However, dynamic content, like the
GeoBox in T162360, make use of the alt attribute so we match the legacy
parser for now and leave making changes as mentioned T297443#7910435 for
the future.

T63566 suggests that we should perhaps be falling back to descriptions
stored with the images, like structured data on commons, instead.

Bug: T297443
Bug: T162360
Bug: T63566
Change-Id: I9d82c8003fd67cd984a8f4523e4993ed1f22b5d2