Account for format being in gallery media options
If a format option is used, the gallery line will be parsed with that
subtype, since it reuses wikilink parsing.  Strip that when copying the
rdfa typeof.

Other media options might not make sense for galleries either, so we
should factor out some reusable portion of the parsing, rather than
stitching together wikilink as we currently do.  T252246 is the same
idea but for serialization.

Bug: T214601
Change-Id: I7e563ca315015f9c26a21bf6566a6d2bdf073629