Internationalization support for Parsoid
This patch is a preliminary patch for the full redlink support in
Parsoid. We introduce a stronger internationalization support in Parsoid
(to be documented in DOM spec 2.6.0) that will allow for a later
localization pass based on the information stored in the i18n data
structure.
In this patch, we also keep the i18n pass, which replaces content of any
mw:I18n span with the string Error: {error message key} - the
localization of this message will be handled in a later patch.

Bug: T309024
Change-Id: Idcd79a85068e082486941b9eb9348aa216851405
Depends-On: I6df964feb09db9acca95b2f32577d69df560d7a6