Implement relative redirects in parsoid extension.
The logic for doing redirects to other endpoints in a "relative" way
is being removed from the ParsoidHandler base class in core, but it's
still needed by the parsoid extension endpoints.

This patch is forward-compatible, it needs to be merged before the core
patch.

Needed by I8e154ded4b93b1e7201fc302a968bd88211666bc.

Bug: T311867
Change-Id: I85ac2f2a8626989e111265869ff301fc72dfaa15