e2e test: cover etag handling with structure test
This replaces the fixed expectation for the page endpoint to return an
etag with a structure test that ensures that any etag returned by the
page endpoint is accepted by the transform endpoint, but it's also ok
for the page endpoint to not return an etag.

Background: both the implementation of the page endpoint and the
handling of etags in the transform endpoint are in flux, and are
impacted by changes in core. To avoid breaking tests in the parsoid
extension, the tests should be relaxed to only assert the contract
required, rather than a concrete behavior.

For reference:
* old behavior: the page endpoint returns an etag that will never match
anything. The transform endpoint ignores the if-match header.
* desired behavior: the page endpoint returns an etag that references a
stored copy of the output it returns. The transform endpoint looks up
cached content based on any etag it receives, and fails with status 412
if it doesn't find a match.
* transient behavior, per Ida81a314f0: the transform endpoint handles
etags and fails if they don't match (see above), but the page endpoint
doesn't return any etags for now.

Bug: T311819
Change-Id: I26059591f8cadf112d22afbe702817f781943198