Update baseconfigs
This is needed so that so that the media options can be used in a follow
up.

However, T91154 added {{=}} as a built-in magic word and necessitates a
native parser function implementation for it for the tests to continue
passing.

We can consider treating in a similar manner as {{!}} in the
TemplateHandler in a follow up.

Change-Id: Ieac2166db897311ec05fa3c443e89a908bfd6b39