Ensure that broken parser functions are still recognized as parser functions
Previously invocations starting with `#` would try to get parsed as
templates, which led to visual editor thinking they referenced a
fragment on the current page (!).

Note: because Parsoid now generates for (broken) parser function `$f`
a title for internal use of `Special:ParserFunction/$f` we need to
update a testing mock slightly to allow it to handle canonicalization
of titles in the `Special` namespace.

One entry added to selser blacklist because selser moves some
whitespace around.

Bug: T314524
Change-Id: Id8be7301863a90792b71a34ce52ae42575f0114a