Rich attribute support, phase 1a
This patch begins to provide a uniform interface to HTML attributes
which contained structured data.  It allows serialization as JSON and
deserialization/use as a proper PHP object type.  This patch
corresponds to "phase 1a" of the proposal at:
https://www.mediawiki.org/wiki/Parsoid/MediaWiki_DOM_spec/Rich_Attributes

This separates the concept of "JSON-valued attribute" from the page
bundle representation: not all JSON-valued attributes are represented
in the page bundle, nor the converse.  It also moves toward a richer
concept of attribute values which will include document fragments in
the next part of phase 1.

Bug: T339927
Bug: T348160
Depends-On: I686390f8c655359ba0a127c76a1f242dba21b9f2
Depends-On: I6087d896bd9f6a4e7375647253dbe142389b8194
Change-Id: I2983d27d40d7c5951f17d62a8f3dd8b6ac6abdef