Add nofollow support to SiteConfig & disable test checks for nofollow
This is a preliminary patch before adding nofollow noreferrer noopener
attributes to the external link rel attributes. We expose the usage of
$wgNoFollowLinks and $wgExternalLinkTarget configuration options in
Parsoid code. We also temporarily disable the check for the values
nofollow noreferrer noopener in Parsoid tests unless testing explicitly
for it - as used in the following tag in this chain.

Bug: T186241
Change-Id: I12a815f19c82212759f399c605650679bb29478a