Restore cleanupFormattingTagFixup pass
It was removed in d473791.  In review of that patch, it's stated,

> that pass can be dropped entirely because we no longer create the
> mistakes in the line based p-wrapper that necessitated it

Yes, the switch to DOM based p-wrapping meant that we avoided all the
mistakes that we were making, so this was much less common.

However, it's still possible for content model violations to induce
reopening of active formatting tags in the media structure during
treebuilding, which have nothing to do with p-wrapping.  In d8564b0,
we noted a few violations that originate from wikitext source.  We
should consider if we want to expose these as lints for editors to clean
up.

The deleted logic from the cleanupFormattingTagFixup pass is
incorporated into the AddMediaInfo pass since we're already doing work
verifying the structure there.

Bug: T314059
Change-Id: I3c55eb5fb8055016f8c4f76d27d953f65ff621be