Make "update-unexpected" option functional in parserTests
This patch refreshes the usage of the "update-unexpected" option in
parserTests.
update-unexpected and update-tests are now both boolean options, which
can be used in conjunction with the new update-format option to provide
an update format.
The "update-format" option defaults to "noDsr", which corresponds to a
raw format, from which dsr is stripped from data-parsoid, matching
what's typically done manually.

Change-Id: Ib378a58a385e6e2d69088bb3f7953428fdcb2d9a