Bump minimum required PHP version to 7.4 in phan
The *target* version is set by mediawiki-phan-config, and is currently
8.1; this checks for forward-compatibility issues.  We need to override
the *minimum* target version, which hasn't (yet) been bumped to 7.4.

PhanPluginDuplicateExpressionAssignmentOperation now wants us to use
the ??= operator where possible.  There were only 6 of these locations,
so I went ahead and updated them.  Shouldn't be too annoying if we
need to backport to 1.38 (1.39 is also PHP 7.4+).

Change-Id: I16c1a976a7ac602d99d5c512aaf663eead7e4aa7