Add methods to convert to/from BCP-47 and MediaWiki-internal language codes
These are used in follow up patches which try to disentangle our use
of these language codes.  The new wikimedia/bcp-47-code library is used
to allow core to pass an opaque language object to Parsoid which can
be converted to a BCP-47 string but hides all its other implementation
details.

(Adds the wikimedia/bcp-47-code library to Parsoid's dependencies, which
requires this patch to Depend-On a corresponding patch to mediawiki-vendor.)

Depends-On: I993098e671ada96d91f6cdb014f83957a306a9ba
Depends-On: Ibca066d0b95c318187323ae76b0e20f28096f6f9
Change-Id: Ica7e96308708e38effc34abbcaf3dc96adb818fd