Split methods to take/return BCP-47 language codes
In order to clarify where we are accepting/returning BCP-47 language
codes versus MediaWiki internal codes, split off *Bcp47() versions of
methods dealing with language codes. This clarifies the status quo
behavior (where MW-internal codes are used in most places) and allows
a gradual transition away from MW-internal codes toward the use of
BCP-47 codes.  (In general we can't just broaden the parameter types
in the interfaces to allow the new Bcp47Code class, because then the
implementation of those interfaces in mediawiki-core would break.
Instead we'll do this in steps: define the new interfaces, update
mediawiki-core to use the new interfaces, remove the old interfaces,
rename the new interfaces to the old names.)

For parameters in options arrays we have it a bit easier, since we
don't have the subclassing restrictions to worry about.  Here we just
start accepting Bcp47Code objects as well as MediaWiki-internal code
strings.  A subsequent patch to mediawiki-core will pass its internal
Language objects here (which will implement Bcp47Code) instead of
passing Language::getCode() and exposing the MW-internal code strings
directly.

The ultimate goal is for MW core to be the sole source of truth regarding
"internal language codes" and Parsoid to have a narrow interface to
an opaque object which lets it convert language objects to BCP 47 strings
(since these appear in HTML) but not otherwise be exposed to them.

Depends-On: I2915e3bc288f4293e4ebf11ab68ccd6d020f2b8e
Change-Id: Ia91d2bffde745c31e285ad8394ca457f825c7ab5