Remove "href missing in a" warning
The introduction of normalization of extension content introduced a
large number of "href missing in a" warnings, due to these hrefs being
sometimes removed by clients in extension content (e.g. T194081).

Removing the warning in "fromEncapsulatedContent" contexts was
considered; but then nothing in the HTML spec prevents the addition of
an a tag without an href; additionally, that warning is not actionable.
Hence, the current patch completely removes it, and returns the node
without further processing (that would anyway not happen with an empty
string href) if there's no href.

Bug: T321281
Change-Id: I799959051f166ec4ca666690c291496c1f953cb0