Fix sealed fragment duplication on node cloning
When a node is cloned during tree builder fixup, if there's a sealed
fragment in the attributes of the node we're duplicating, it gets cloned
as well, leading to issues when we try to resolve it down the line. This
patch removes the fragment duplication in that case.

The newly introduced tests crash without this fix.

This is only a first step in addressing T260082: the behaviour is
different compared to unsealed fragment (in which case the fragment
is not stripped).

Bug: T260082
Change-Id: I6434174416c0cfd7320192670ead627b9e3ae851