Rename SectionMetaData -> SectionMetadata; other improvements
Use ?string for the `fromTitle` property, as it can be missing in
some cases.  Convert this to/from `false` to be consistent with the
way sections were serialized by the existing API methods.

Use enumerated arguments for the SectionMetadata constructor, although
preserve a `fromArray()` factory method for ease of conversion of
existing code.  This constructor is a great candidate for named
arguments when our minimum PHP version is >= 8.0.

Change-Id: I2c9dfa6463530aa6b94762dbcff0a5717ce2eabc