Fix wikilink parsing issues on redlinks
If wikilink contains a : followed by a number, this is interpreted by
parse_url as a port number, and the URL is not re-formed correctly. It
can even lead to crashes when subsequently removing the redlink
parameters from the URL. This patch fixes this case.

Follow-up to I3c316b01185c770846598fde00bfd8cb75cfa27d.

Bug: T309024
Change-Id: Ibea703db61500714e0af6b664311a0ba6527d05f