Use Parsoid's UrlUtils::parseUrl to parse wikilinks
Follow up to c2948df75

This also fixes an issue that, when refreshing redlinks with the pb2pb
pass, the querystring params added that no longer apply to a page were
failing to be removed, similar to how the "new" class is cleared.

If that feature is even still useful, we should add a test that makes
pass that was previously missing come into existence before refreshing
the redlinks.

Bug: T325329
Co-Authored-by: Isabelle Hurbain-Palatin <ihurbainpalatin@wikimedia.org>
Change-Id: I995805c1164e134e7e3c5f4edfb3d520e5d06716