TOCData: workaround for non-recursive JsonCodec
Without this change, when serialized in core with the current JsonCodec,
the SectionMetadata objects will fail to serialize since the JsonCodec doesn't
recursively serialize embedded objects.  This is added to JsonCodec in
Id7d0695af40b9801b42a9b82f41e46118da288dc, but this workaround allows
the section refactoring work to be independent of that JsonCodec work.

Bug: T312589
Change-Id: I41be8d02c2aa775cc673c77ac64273e3d189f0c5