Address FIXMEs in WTUtils A-tag helpers
The changes introduce a minor performance penalty but it is probably
not significant in the scheme of things given that this is only
on A-tags.

The code simplification and robustness is probably worth the penalty.

The FIXME related to needing stx to infer the syntactic form for the
A-tag comes from 2014 when I seemed to have thought that we should be
able to infer that information without relying on the stx property.
But, I no longer believe that to be useful -- stx is useful information
derived form source and is as reliable a piece of information as any
about syntactic forms. So, I simply got rid of that FIXME.

Change-Id: I80423e11e23d27848b6751659d86f504906f283d