Remove the paragraph rule
Since f4081be, paragraph wrapping has been done on the token stream and
on the dom.  Since we're no longer emitting p tag token and inlineline
is already the next expression we try to match in block, this no longer
seems to serve a purpose.

This probably also means the distinction of block_tag is no longer
relevant.  It will be explored and removed in a follow up.

Change-Id: Ic50bbe1e932080dfeb21cceff72836bb6bc13351