Remove the block_tag rule
In 5fb2e33, the paragraph rule was removed since it no longer emitted p
tags in the tokenizer, paragraph wrapping being left to the token stream
and the dom.  Therefore, the purpose of distinguishing block_tag in
block, to avoid the unnecessary wrapping, is also no longer needed.

Similarly, in 4b22215, the pre_indent rule was removed since indent pre
is handled in the state machine on the token stream.  The block_tag in
block_line, to avoid triggering indent pre, is no longer in use.

Change-Id: I724889e7b0cb36e23d85f701635d4ec9a4e60086