Add TOCData::getExtensionData() and SectionMetadata::getExtensionData()
Somehow we forgot to add getters when we added the setters in
I5ff0affb7f697c5dbee438217d126736dfd95756 and
I1b267d23cf49d147c5379b914531303744481b68.  The extension data was
included in the legacy array output, and that's how downstream was
accessing them, but we should have 'real' getters.

Change-Id: Ifad819ba83cd1c7105955d24abb1f49268fd2bd1