Use alt as link content for broken media
This is what browsers will do if an image can't be loaded but we're
preempting that with a span, rather than an img tag with a broken src.

It is closer to what the legacy parser does (see the legacyMedia.txt
parserTests) except that it only uses the caption text.  Aligning that
in core is in Ic24fcbb932d02f1c820d4cb1c353d63844a023f6.

Bug: T273014
Depends-On: I7ab3d141b1df92d4447f7e3d6164082844d5bd10
Change-Id: Iddf3e204d6e489cc8a33034da0d9e540efe65553