SectionMetadata: Rename byteOffset to codepointOffset to reflect reality
Contains a temporary SectionMetadata::__get()/__set() implementation
to ensure that code can continue to use ->byteOffset during a transition
period during the deployment of this patch.

The property is still named 'byteOffset' in the "legacy" array, which
ensures that any consumers of the action API (eg) won't break.  But
new code should be able to use the new, correct, property name.

Codesearch doesn't indicate any other users of
SectionMetadata::$byteOffset (which isn't surprising, as it was newly
added in 1.40):
  https://codesearch.wmcloud.org/search/?q=-%3EbyteOffset&i=nope&files=&excludeFiles=&repos=

Bug: T319141
Change-Id: Ide436dca5a609c866da3c63049723243b8242c34