Move default file namespace to gallery
Follow up to I55942bbb36c3bd3dd43a665fbca0f275e9870f7c

The flexibility on providing the file namespace prefix to the title is
special to galleries, in contrast to imagemap and general wikilinks.
Move that handling the gallery extension, along with the hack necessary
to shadow it for roundtripping.

The added imagemap parserTest shows that Parsoid was previously too
lenient compared to the legacy parser, in that it would have rendered
the imagemap and failed the test.

Change-Id: I777698b73f02e9ab10c915b6c75cd2566c2dcb99