Remove $env from DiffUtils
So that these utilities can be used in contexts like extensions without
having to expose the environment there.

The page id check was added in I69806740329a2094ab854577c3f20c7adc4db237
with the idea of using it for an efficient 'git blame'-style authorship
attribution of edited content, which hasn't borne fruit in the past 10
years.

Change-Id: I25d043494f41acdde9dc6840c307557b145cf471