Title.php: Check for null return of preg_replace
This is a port to parsoid of d60cde2ad9a4ed8f25105e1ff53b7cabbbf81245
in core.  Returning a null from preg_replace means an error happened
in processing the regex. If the regex being used is constant and a
unicode one, that means that almost certainly the issue is an invalid
utf-8 sequences. So throw appropriate exception if that happens.

Bug: T313663
Change-Id: I994cf0cc7f31eaed06359fcfdf682ad2a3a91d01