Title.php: remove unreachable parts of title invalid regexp
This is a back-port of 5ad1190f67a26f1a397724d9e783766c14ded2e4 in
core.

In a3a2744d03 (then r28968), the illegal title character regex was
expanded to block XML/HTML character references. However even back
then the two parts of the regex that aimed to block character
references with hashes in them were unreachable, because the fragment
split happened one step before illegal characters were checked. Any
character reference with a # in it would be split across the dbkey and
fragment, and the regex is only run against the dbkey.

Bug: T297578
Change-Id: I156eeb2d02562b5050d87dbbd9e6efad7bba6bb1