Title.php: bring closer to MediaWikiTitleCodec; allow colons in Talk titles
This backports more of the code structure from core's
MediaWikiTitleCodec::splitTitleString(), with (only) two main differences:
Parsoid's implementation doesn't split the interwiki prefix from the
title, and Parsoid's implementation doesn't specially handle interwiki
links which happen to point to the local wiki, the way core does.

In the process of fixing this code, a proper test case for T329690
was written, and the issue with colons in Talk page titles found
in T332903 was fixed (and a test case written).

Bug: T332903
Bug: T329690
Followup-To: I68245fbe7d83d580c8d37a2284678cacd1e0e609
Change-Id: I07e3e32e9f7300feb99427cae9cbc35609c3a886