Fix regenerating old html from cli
Wikitext/ContentModelHandler::toDOM, which is called to regenerate the
old dom, parses from a $pageConfig->getPageMainContent().  So, we need
to set it to the passed in --oldtext(file) content, which is stored in
$selserData->oldText, if we want it to work.

T234549 proposes doing away with storing oldText in SelserData for
consistency with the wt2html direction.  It would avoid such issues but
we need to be careful about not making use of it when we aren't in
$state->selserMode.

Bug: T234549
Change-Id: I22b46c4cf8be13aedefef22c1a046532af9ad667