No need to resolve subpages when resolving lonely fragments
The string isn't referencing a subpage, so exit early.

This fixes the crasher from the task, captured here in the test that
otherwise goes into infinite recursion.

However, there's also probably the case where a subpage is trying to be
resolved on a page with the title "/" that will still hit this bug.  To
be followed up.

Bug: T335847
Change-Id: I5b68889881f982b670f3d531b9bb2eeba7ad8ed5
(cherry picked from commit 18953c47c0ab2203f62c4aa5b8dd0c78e5857df2)