Add a utility to parse extension tag name from typeof
There's an inconsistency in use sites on whether we derive the
extension tag name from the typeof annotation or retrieve it from the
data-mw.  In the parsing direction, they are both set to the same thing.

The spec doesn't say anything about what's expected or required for
serializing.

Noting that we could have gone with,

mw:Extension/cite and data-mw->name="ref" or "references"

instead of,

mw:Extension/ref and mw:Extension/references

But since the latter is what's current, we'll consider that
canonical.

Follows-Up: If921ca471f25f21671a60c3f796fdf145267364d
Change-Id: Ic80cb1593f094527c68c19bf5f5cd0d52001a07b