Fix retokenizing when bailing on links
The trailing bracket should be available to recombine with content since
an opening bracket seems to be valid link content.

An entity in the url position of an extlink will allow it to avoid a
valid protocol check (because the entity will be represented as a token
array and the protocol check assumes that arrays come from templates but
which can probably be tightened up) until it reaches the extlink
handler, post-template expansion.  At which point, an invalid protocol
will force the extlink to bail.

Bug: T340192
Change-Id: I7815a1e4b04516eeacac1502590b86cd0159eb5f