Enable processing of attribute-embedded-HTML in some DOM passes
The changes in parser tests known failures shows the cleanup pass
in action in attribute-embedded HTML.

In order to avoid trying to store embedded docs multiple times,
I had to suppress calling saveDataParsoid on the embedded docs.
Strictly speaking, this is not a regression, but as the FIXME documents,
this is probably broken behavior, but will need us to write a custom
DOm traversal code to ensure we don't try to store the same doc twice.

Bug: T214994
Change-Id: Idd25d6fde979bd32b1a1de2211c9f4fdb8f6d9cd