Rename processAttributeEmbeddedHTML in DOMTraverser
I find the usage of processAttributeEmbeddedHTML to indicate both
that a pass should be applied (in DOMTraverser) and where it should
be applied (in ParsoidExtensionAPI) to be confusing. This patch
suggests an alternative, that (possibly after bikeshedding)
ideally would be used in Idd25d6fde979bd32b1a1de2211c9f4fdb8f6d9cd
to name the processor option too.

Change-Id: I4868de541761900d020af514b183e6dc9ff0f9ed