Update ContentMetadataCollectorCompat to allow RT testing
This adds a no-op stub implementation of the `appendOutputStrings()`
method recently added to ContentMetadataCollector.  Even though the
corresponding patch to add an implementation to ParserOutput has
already landed in core, the "old" version of core being used for
Round-Trip testing doesn't yet have this method implemented.  This
trait allows us to offer a stopgap.

Remove the old TOC-related stub which was added for a similar
purpose but is no longer necessary.

Change-Id: Ibda8b179dac93c37ecb6ed9ec2b1acbf6d5789e8