Introduce SiteConfig::getLinterConfig() and pull in all linter configuration
This patch also tweaks the method signature for SiteConfig::linting() to
allow fine-grained per-lint enable/disable.  No one actually overrides
SiteConfig::linting() in client code, so this signature should be safe
to change.

Deprecate the large-table related parameters in SiteConfig and replace them
with a general `SiteConfig::getLinterConfig()` method, which could be sourced
from a configuration variable in core.

Bug: T343314
Followup-To: I9249de797da0f412d659d54b55c49e5d2238fbab
Change-Id: If538fd937d97b63a40fd72e62c93bcb1b3224366