Ignore metadata tags in textContentFromCaption
The method is used to populate the content of title and alt attributes
on media and the presence of templatestyles in captions confounds it.

A test is added that captures the semantic failure from roundtrip
testing.  In order to avoid serializing an alt media option
unnecessarily, a comparison of the text content of the caption and alt
text is made and, if equal, the alt is suppressed since if there isn't
an explicit alt, it'll be set to the caption text when parsing.
However, prior to this patch, the style content from the templatestyle
was included in the alt yet, because of deduplication of templatestyles,
the style tag in the caption later becomes link tag without any text
content and then the comparison would fail.

Bug: T339237
Change-Id: Ice62d61d1a4f661770869b43d54eb8fc20824b71