Remove PageConfig::hasLintableContentModel
This has been superseded by ParsoidOutputAccess::supportsContentModel
in core which checks if the content model serializes to wikitext or if
Parsoid has a native content model handler, like for JSON.

A method that checks if a content model serializes to wikitext may be
useful more broadly but we wouldn't call it hasLintableContentModel and
it would be used outside the linter to suppress parsing with Parsoid
altogether.  Parsoid parses wikitext which is lintable.  The
hasLintableContentModel method was introduced because Parsoid was being
asked parse to various content models, like JavaScript, as wikitext
which would inevitably produce unhelpful linting errors to sift through.

Change-Id: I04d102c4deb0d72b4c85f41a2e8500d0cb13188b