Magic words aliases aren't necessarily unique to a magic word
Both "img_sub" and "sub" have the alias "sub" so the magicWordMap has
collisions.  Here we create a separate media option map that only
contains the aliases for media options.  This is akin to core where the
magicWordFactory is used contextually.

This probably wasn't noticed earlier because the order in which magic
words are added to the map differs on dewikivoyage.

Bug: T345026
Change-Id: Id16d087bd24cc8f2fa1fea90b992441a0b28edbb