Remove infinite loop in AnnotationDOMRangeBuilder, take 2
Follow-up on I040eb3a37e168562fcc1ab0ea3948f3d7a332f1c, which was
removing the infinite loop but keeping the conditions that was making
the annotation wrapper not necessarily return balanced trees between
annotations because of the "uneditable" wrapper insertion.
This patch ensures that both ends of an annotation range have the same
parent.

Bug: T325575
Change-Id: I5fc4229c5ae1dc0d4890967b428dba26b1dc013a