mw:DOMFragments shouldn't copy inTransclusion flag from first encap node
Another case similar to I4f7a6392581f969d8293e39fd825ffddc314c003
where flags set on the first encapsulation node are wrongfully being
transferred to the mw:DOMFragment token, interacting with a different
pipeline.

Template wrapping is local to a pipeline (skipNested is false in the
post-processor) and so the inTransclusion flags that are set when tree
building should be cleared at the end of a pipeline.  That mostly
doesn't matter except for mw:DOMFragment tokens which sometimes get a
copy of the data-parsoid attribute from the first node they are
encapsulating.

Everything that follows is just as a result of false assumptions once
that flag is left around to be used in another pipeline.

When treebuilding, inTransclusion flags are set on elements so that if
they're found in a foster box, we can wrap the entire table and foster
box in template meta markers.  Then, when template wrapping, the
algorithm looks at the overlapping ranges and transfers the subsumed
range info to the compound template range info.

For mw:DOMFragments, template wrapping has already occurred in the
nested pipeline.  If the flag hanging around on the first encap node is
transferred to the dom fragment node and the node is fostered, it will
induce the foster box to be template  marker wrapped.  However, there's
no subsumed range to be found because the flag is from a different
pipeline.  Subsequently, no template info gets set for the range and we
get a crash.  Meta markers added around the foster box don't have
template info of their own and rely on info being propagated up.

It's worth noting that if an mw:DOMFragment token needs an
inTransclusion flag in its current pipeline, that'll get set during
treebuilding.

The added test crashes without the fix.

Over the last 30 days there have been ~500 occurrences of this crasher.
Checking the latest handful, this fixes,

php bin/parse.php --oldid 173490 --domain ja.wikinews.org < /dev/null
php bin/parse.php --oldid 1160769070 < /dev/null
php bin/parse.php --oldid 31686287 --domain pt.wikipedia.org < /dev/null
php bin/parse.php --oldid 35108915 --domain uk.wikipedia.org < /dev/null
...

Bug: T345705
Change-Id: I5d120b7b4123d6ed4ef14b93a40fbbc18c166fd3