Introduce stripping of annotation in non-wikitext extension content
This patch introduces a new "stripMethod" configuration for annotations
that aims at providing a callback so that extensions defining
annotations can strip their markup from a given string.

It also introduces a 'hasWikitextInput' configuration option for
extensions so that extensions can opt into some preprocessing. The only
preprocessing right now is to strip the annotation markup.

Bug: T341009
Change-Id: I4e9a7a8bec3cb9532ef8a729fd2c6c4acca5d8a0