Putting fostered end annotation markers after the table
If an annotation end marker gets fostered, it can happen that it ends up
before its corresponding start marker in the tree. We want to avoid
that, and instead move the end marker after the fostered table, so that
the whole table is marked as uneditable because the range is extended.
This patch handles that case and provides a test case that would fail
without the patch in MarkFosteredContent.

This behaviour differs from the template range detection (which handles
flipped ranges) for various reasons:
- we can get dangling annotation end markers (whereas we cannot get
  dangling template end markers), which we want to ignore, but we cannot
  ignore them if they are not guaranteed to be after their start tag
- we want to mark a whole table as uneditable if an annotation tag ends
  in a fosterable position, so we have to handle fosterable position
  explicitly anyway.

Bug: T347180
Change-Id: If1b76ceb7383318a4cd2c728b8f3b0a1fd74b085