Include fostered annotation tags in fostered transclusions
Follow-up to If1b76ceb7383318a4cd2c728b8f3b0a1fd74b085.

When there's a fostered transclusion, the whole table gets wrapped into
a transclusion. We need to adjust the position of the annotation metas
that get moved to "after the table" so that they still get included in
the transclusion, which is what this patch does.

The introduced test tests this by testing wt2wt, which shows another
issue (not fostering-related), documented in T347877; the wt2wt test
demonstrates that the end tags get properly wrapped in the transclusion.

Bug: T347180
Change-Id: If9219c34bd0c6548ef9c2eef0b02b21f8672a362