Annotations in template targets
I487baaafcf1ffd771cb6a9e7dd4fb76d6387e412 was about annotations in
template arguments and it was left to the Translate extension to strip
the translate tags before preprocessing for the time being.

However, annotations in the template target are causing the template
target to fail to resolve before it can be sent off for preprocessing.
Here they are ignored for the sake of stringifying the target, to avoid
bailing on the template altogether.

The probably revives the larger discussion about where the stripping
should happen, who's responsible, and whether it's a property of
annotations in general or just the translate tags.

The test is added to the knownFailures list because the patch in
Ia36f97f9ed92bc252f76fd63859b0aff1e7ec382 fails to account for if the
annotation is template marked before stringifying it.  Presumably the
fix is the check for an about marker and span wrap the text node if so.
But, that will be in a follow up.

Bug: T349078
Change-Id: Iba424d49093ae98b7bd482090c58b07466704fdc