Reuse LinkHandlerUtils::figureHandler in gallery
Adds a ParsoidExtensionAPI::serilalizeMedia that's intended to act as
the converse to ::renderMedia.

Various media options (like "thumbtime", "page", and "lang") were being
lost in galleries during roundtripping because, although they were
supported in parsing, the serialization code was a poor duplication of
all the complexity of the figureHandler.  Drying out the media
serializations fixes the issues (as seen in all the newly passing
tests) and makes it more maintainable going forward.

Changes to tests:

"Gallery override link with wikilink (T36852)"

The order of the media options is switched to avoid having to add
data-parsoid sections to preserve the order.  ::figureHandler serializes
them in a different default order than gallery's ::contentHandler used
to.

"Gallery override link with absolute external link (T36852)"

A data-parsoid attribute is added to preserve the order of options when
serializing.  Swapping the options doesn't work because the legacy
parser would consider the "|" as part of the link url.  Pretty sure
that's a known difference.

"Gallery with invalid title as link (T45964)"

A data-parsoid attribute is added to match the test,
"Image: Invalid title as link".  In both cases, Parsoid seems to produce
a dirty diff when serializing from data-mw->attribs.  Needs to be
investigated separately.

"Selser preserves unedited gallery lines"

Again, just a difference in default serialization ordering of options.

Bug: T252246
Bug: T214603
Bug: T333907
Change-Id: I7a5c24f6ffc15ff0455adf5b025b695ee71501b6