Remove variant-metadata-hack; add helper to add http-equiv meta tags
The DOMUtils helper can let core use it to add meta tags to the header
But, it still needs knowledge of what headers to add for the variant
transform where the transform comes from core. So, the cost of removing
the variant-metadata-hack pb2pb transform is that the knowledge of what
metadata to add exists both here and in core and would have to be kept
in sync.

Change-Id: I0a41933391eac5dcad5de59bd18a09ac4fcc9037