Don't use count() when the actual count is not needed
While count() is not a slow function any more in modern PHP, I still
find it odd to see it in code that doesn't care about the actual
count. What this code cares about is:
* If an array isn't empty. We can use `$array !== []` or even
  `if ( $array )` instead.
* Just iterate the array with foreach.

I tried to focus on usages of count() that showed up as hot in an
xdebug profile run.

Change-Id: I1a8c3ea4b54779bc938843a9cb91ca6ed88742ef