Optimize hot SiteConfig::namespaceId
I notized this is an extremely hot method when running the PHPUnit
test suite for this codebase. It might be that this is only because
of the test setup and not that relevant in production. Still I find
it worth unrolling this code a bit. Personally I find this more
readable.

* Avoid calling the normalization twice.
* Generally avoid an extra level of function call overhead.

Change-Id: Ibc3aaac48b28677a0336cb78c710d3a128b4f14a