Optimize very slow SiteConfigTest
This reduces the total runtime of SiteConfigTest from 600ms (about
1/4 of the total runtime of the test suite) to effectively zero.

I'm not entirely sure what the purpose of the "too many" aspect of
this test was. I looked around but couldn't find any code that would
do anything special when there are to many entries in the array, or
to much memory is consumed. It seems the actual number doesn't matter
that much and can as well be something else.

There are 3 test cases:
1. The "x76" is a match within the range.
2. The ":l" matches twice. One time within the range and a second
   time with the dedicated "l" entry that's appended to the end. The
   expectation is that the later one matches, which it does.
3. The "9999999" was and still is outside of the range.

Change-Id: I1de88cd722f447baee40308eb61b0c1ba548b0df