Add fast helper method DOMUtils::hasClass
While this is not really a bottleneck (usually the list of classes
is short) I was still curious. Why create an object, explode a
string into an array (even using a regex) and search the array if
we can do the same with a single regex?

This will be beneficial for other codebases, notably
DiscussionTools.

Change-Id: If9252a97562542e7a4bec29edcc6e8dee0fb8221