Emit relative redirects.
Using relative paths in the Location header of redirect responses avoids
issues when the API is called internally in the service mesh.

This patch removes two methods that were overriding methods from
ParsoidHandler in core, to address the issue described in T311867.
However, per Ibb38c30eb66a, the methods in core don't have that
issue anymore, and they are actually more robust, also avoiding
T350219. Removing the overrides allows the parsoid endpoints to
benefit from the new logic in core, namely, to use only the path
in the Location header, and omit protocol, host and port.

Bug: T349001
Bug: T350219
Bug: T311867
Change-Id: I9b7b69c4ddf76f6bcea3a725aafcec9e008ed4c6