Rewrite DOMUtils::nodeDepth() for speed
This is rarely called and not critical for the overal performance in
this codebase. Still I find it worth it. It's a good example for the
kind of performance optimization to look for: No need to create and
discard an array when all we need in the end is a number.

The new implementation needs about 1/4 of the original runtime.

Change-Id: Id998651577e34ddecb0cb8c8db1757f66982cfd7